refactor: 테스트 파일도 올바른 패키지 위치로 이동
