+++ src/test/java/wooteco/subway/domain/fare/DiscounterTest.java
@@ -1,9 +1,8 @@
-package wooteco.subway.domain;
+package wooteco.subway.domain.fare;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import wooteco.subway.domain.fare.Discounter;
 
 import static org.assertj.core.api.Assertions.assertThat;
